# add hash benchmarking to performance tests