I tweaked the performance tests to generate just a ruby hash using both the libraries. This helps demonstrate the performance difference without any involvement of OJ. As OJ is only used in both libraries to convert the ruby hash to a JSON string.

Here is the benchmark output i get when i run these tests

```
Serialize to JSON string 1 records
Serializer      Records    Time
AMS serializer  1          3.21 ms
Fast serializer 1          0.17 ms

Serialize to Ruby Hash 1 records
Serializer      Records    Time
AMS serializer  1          0.7 ms
Fast serializer 1          0.04 ms
.
Serialize to JSON string 25 records
Serializer      Records    Time
AMS serializer  25         16.35 ms
Fast serializer 25         0.6 ms

Serialize to Ruby Hash 25 records
Serializer      Records    Time
AMS serializer  25         10.25 ms
Fast serializer 25         0.28 ms
.
Serialize to JSON string 250 records
Serializer      Records    Time
AMS serializer  250        143.51 ms
Fast serializer 250        3.44 ms

Serialize to Ruby Hash 250 records
Serializer      Records    Time
AMS serializer  250        98.55 ms
Fast serializer 250        2.17 ms
.
Serialize to JSON string 1000 records
Serializer      Records    Time
AMS serializer  1000       556.1 ms
Fast serializer 1000       14.3 ms

Serialize to Ruby Hash 1000 records
Serializer      Records    Time
AMS serializer  1000       367.52 ms
Fast serializer 1000       9.35 ms
.
Serialize to JSON string 1 with includes and meta
Serializer      Records    Time
AMS serializer  1          1.85 ms
Fast serializer 1          0.13 ms

Serialize to Ruby Hash 1 with includes and meta
Serializer      Records    Time
AMS serializer  1          1.2 ms
Fast serializer 1          0.05 ms
.
Serialize to JSON string 25 with includes and meta
Serializer      Records    Time
AMS serializer  25         17.02 ms
Fast serializer 25         0.49 ms

Serialize to Ruby Hash 25 with includes and meta
Serializer      Records    Time
AMS serializer  25         11.98 ms
Fast serializer 25         2.11 ms
.
Serialize to JSON string 250 with includes and meta
Serializer      Records    Time
AMS serializer  250        159.24 ms
Fast serializer 250        7.65 ms

Serialize to Ruby Hash 250 with includes and meta
Serializer      Records    Time
AMS serializer  250        119.94 ms
Fast serializer 250        3.87 ms
.
Serialize to JSON string 1000 with includes and meta
Serializer      Records    Time
AMS serializer  1000       623.21 ms
Fast serializer 1000       30.67 ms

Serialize to Ruby Hash 1000 with includes and meta
Serializer      Records    Time
AMS serializer  1000       458.59 ms
Fast serializer 1000       27.64 ms

```